--- enterprise/frontend/src/metabase-enterprise/data-studio/metrics/components/MetricHeader/CardMoreMenu/index.ts
@@ -1,1 +1,0 @@
-export * from "./CardMoreMenu";
