+++ enterprise/frontend/src/metabase-enterprise/data-studio/metrics/components/MetricHeader/MetricMoreMenu/index.ts
@@ -1,0 +1,1 @@
+export * from "./MetricMoreMenu";
