+++ enterprise/frontend/src/metabase-enterprise/data-studio/metrics/components/MetricHeader/MetricTabs/index.ts
@@ -1,0 +1,1 @@
+export * from "./MetricTabs";
