--- enterprise/frontend/src/metabase-enterprise/data-studio/metrics/pages/MetricOverviewPage/CardOverview/index.ts
@@ -1,1 +1,0 @@
-export * from "./CardOverview";
