+++ enterprise/frontend/src/metabase-enterprise/data-studio/metrics/pages/MetricOverviewPage/MetricOverview/DescriptionSection/DescriptionSection.tsx
