+++ enterprise/frontend/src/metabase-enterprise/data-studio/metrics/pages/MetricOverviewPage/MetricOverview/index.ts
@@ -1,0 +1,1 @@
+export * from "./MetricOverview";
