+++ frontend/src/metabase/collections/components/CollectionEmptyState/CollectionEmptyState.tsx
@@ -87,7 +87,7 @@ function getDefaultEmptyStateMessages(collection: Collection | undefined) {
   switch (PLUGIN_DATA_STUDIO.getLibraryCollectionType(collection?.type)) {
     case "models":
       return {
-        title: t`No tables or models yet`,
+        title: t`No published tables yet`,
         description: t`Publish tables in the Library to see them here.`,
       };
     case "metrics":
